+++ packages/api/lib/assertions.js
@@ -6,12 +6,12 @@ exports.isAuthorizationMissingResponse = (t, response) => {
 };
 
 exports.isInvalidAccessTokenResponse = (t, response) => {
-  t.is(response.status, 403);
+  t.is(response.status, 401);
   t.is(response.body.message, 'Invalid access token');
 };
 
 exports.isExpiredAccessTokenResponse = (t, response) => {
-  t.is(response.statusCode, 403);
+  t.is(response.statusCode, 401);
 
   t.is(response.headers['Content-Type'], 'application/json');