+++ packages/api/tests/endpoints/collections/update-collection.js
@@ -79,7 +79,7 @@ test('CUMULUS-912 PUT with pathParameters and with an invalid access token retur
     .put('/collections/asdf/asdf')
     .set('Accept', 'application/json')
     .set('Authorization', 'Bearer ThisIsAnInvalidAuthorizationToken')
-    .expect(403);
+    .expect(401);
 
   assertions.isInvalidAccessTokenResponse(t, response);
 });